Remove migration logic for controller
